Updated Scripts for Ala10 and NTL9
